+++ src/puyoPage/pages/gamePage/puyoGameConstant.hpp
@@ -14,14 +14,14 @@ namespace puyoGameConstant
 
     const int BOARD_FALL_GRAVITY_TICK = 310;
     const int BOARD_BASIC_VANISH_TICK = 1500;
-    const double VANISH_SOON_PROP = 0.85, VANISH_STAY_PROP = 0.7;
+    const float VANISH_SOON_PROP = 0.85, VANISH_STAY_PROP = 0.7;
     const int BOARD_OBSTRUCT_VANISH_TICK = 600;
     const int BOARD_FLY_TICK = 1670;
 
-    const double PLAYPUYO_IN_BOARD_SPAWN_X = 2;
-    const double PLAYPUYO_IN_BOARD_SPAWN_Y = -0.9;
+    const float PLAYPUYO_IN_BOARD_SPAWN_X = 2;
+    const float PLAYPUYO_IN_BOARD_SPAWN_Y = -0.9;
 
-    const double OBSTRUCT_PUYO_SPAWN_Y = -2;
+    const float OBSTRUCT_PUYO_SPAWN_Y = -2;
 
     const int OBSTRUCT_PUYO_MANY = 18;
     const int OBSTRUCT_PUYO_MID = 6;
