+++ src/puyoPage/pages/gamePage/puyoPuyo/puyoAction/puyoPuyoTurn.hpp
@@ -18,14 +18,14 @@ class puyoPuyoTurn : public puyoPuyoAct
             LEFT
         };
         Direction dir;
-        const double rad, c, s;
+        const float rad, c, s;
         puyoPuyo& center;
 
         std::array<std::unique_ptr<puyoPuyoAct>,2> sub_acts;
         bool test(puyoBoard& board, puyoPuyo& puyo) override;
         void arrive(puyoPuyo& puyo) override;
     public :
-        puyoPuyoTurn(int amount, puyoPuyo& center, std::pair<double,double> turning);
+        puyoPuyoTurn(int amount, puyoPuyo& center, std::pair<float,float> turning);
         bool decline(puyoBoard& board,puyoPuyo& puyo) override;
         void act(puyoPuyo& puyo) override;
 };
